+++ models/vicuna.py
@@ -8,7 +8,7 @@ __all__ = ["Vicuna15"]
 
 
 class Vicuna15(Llama2):
-    def __init__(self, vicuna_15_path="lmsys/vicuna-7b-v1.5", dtype=torch.float32):
+    def __init__(self, vicuna_15_path="lmsys/vicuna-7b-v1.5", dtype=torch.float16):
         model = AutoModelForCausalLM.from_pretrained(
             vicuna_15_path, torch_dtype=dtype, trust_remote_code=True, use_auth_token=True, top_p=None
         ).eval()
